finished test

+++ bookkeeper-server/src/test/java/org/apache/bookkeeper/client/BookKeeperAdminInitBookieTest.java
@@ -1,14 +1,12 @@
 package org.apache.bookkeeper.client;
 
-import org.apache.bookkeeper.bookie.BookieException;
+
+import com.google.common.util.concurrent.UncheckedExecutionException;
 import org.apache.bookkeeper.bookie.BookieImpl;
-import org.apache.bookkeeper.bookie.storage.ldb.DbLedgerStorage;
-import org.apache.bookkeeper.common.allocator.PoolingPolicy;
+
 import org.apache.bookkeeper.conf.BookKeeperClusterTestCase;
 import org.apache.bookkeeper.conf.ServerConfiguration;
-import org.apache.bookkeeper.discover.RegistrationManager;
 import org.apache.bookkeeper.meta.zk.ZKMetadataDriverBase;
-import org.apache.bookkeeper.net.BookieId;
 import org.apache.bookkeeper.util.BookKeeperConstants;
 import org.apache.commons.io.FileUtils;
 import org.junit.After;
@@ -22,16 +20,17 @@ import java.io.File;
 import java.util.Arrays;
 import java.util.Collection;
 
-import static org.apache.bookkeeper.meta.MetadataDrivers.runFunctionWithRegistrationManager;
 
 
-/**
- * Test the bookkeeperAdmin method "initNewCluster()"
- */
+
 @RunWith(value = Parameterized.class)
 public class BookKeeperAdminInitBookieTest extends BookKeeperClusterTestCase{
 
-    boolean expectedResult;
+    private boolean expectedResult;
+
+    private final boolean deleteDirs;
+
+
     private static final int numOfBookies = 2;
     private final boolean useServerConfiguration;
     private final int bookieindex = 0;
@@ -39,7 +38,11 @@ public class BookKeeperAdminInitBookieTest extends BookKeeperClusterTestCase{
 
     @Before
     public void setUp() throws Exception {
-        super.setUp();
+        try {
+            super.setUp();
+        }catch (Exception e){
+            e.printStackTrace();
+        }
         //kill a potential existing bookie
         if (useServerConfiguration) sv = confByIndex(bookieindex);
 
@@ -49,23 +52,23 @@ public class BookKeeperAdminInitBookieTest extends BookKeeperClusterTestCase{
         }catch (IndexOutOfBoundsException e){
             System.out.println("there is no bookie to kill");
         }
+        if (sv!=null && this.deleteDirs) {
 
-        File[] journalDirs = sv.getJournalDirs();
-        for (File journalDir : journalDirs) {
-            FileUtils.deleteDirectory(journalDir);
-        }
-
-        File[] ledgerDirs = sv.getLedgerDirs();
-        for (File ledgerDir : ledgerDirs) {
-            FileUtils.deleteDirectory(ledgerDir);
-        }
-
-
-        File[] indexDirs = sv.getIndexDirs();
-        if (indexDirs != null) {
-            for (File indexDir : indexDirs) {
-                FileUtils.deleteDirectory(indexDir);
+            File[] journalDirs = sv.getJournalDirs();
+            for (File journalDir : journalDirs) {
+                FileUtils.deleteDirectory(journalDir);
             }
+            File[] ledgerDirs = sv.getLedgerDirs();
+            for (File ledgerDir : ledgerDirs) {
+                FileUtils.deleteDirectory(ledgerDir);
+            }
+            File[] indexDirs = sv.getIndexDirs();
+            if (indexDirs != null) {
+                for (File indexDir : indexDirs) {
+                    FileUtils.deleteDirectory(indexDir);
+                }
+            }
+
         }
 
 
@@ -82,6 +85,7 @@ public class BookKeeperAdminInitBookieTest extends BookKeeperClusterTestCase{
 
         //kill the created bookie
         killBookie(bookieindex);
+        super.tearDown();
 
 
 
@@ -90,35 +94,48 @@ public class BookKeeperAdminInitBookieTest extends BookKeeperClusterTestCase{
 
 
 
-    //Parametri in input
+    //input parameters
     @Parameterized.Parameters
     public static Collection<?> getParameters(){
         return Arrays.asList(new Object[][] {
-                //expected result, useServerConfiguration
-                {true, true}
-                //{true, data, 2, data.length-2, true, sync},
+                //expected result, useServerConfiguration, deleteDirs
+                {true, true, true},
+                {false, false, true},
+                {false, true, false},
+
         });
     }
 
-    public BookKeeperAdminInitBookieTest(boolean expectedResult, boolean useServerConfiguration) {
+    public BookKeeperAdminInitBookieTest(boolean expectedResult, boolean useServerConfiguration, boolean deleteDirs) {
         super(numOfBookies);
         this.useServerConfiguration = useServerConfiguration;
         this.expectedResult = expectedResult;
+        this.deleteDirs = deleteDirs;
 
     }
 
     @Test
     public void initBookie() throws Exception {
         System.out.println("starting test");
+        boolean result = false;
 
-        String bookieId = BookieImpl.getBookieId(sv).toString();
-        String bookieCookiePath =
-                ZKMetadataDriverBase.resolveZkLedgersRootPath(sv)
-                        + "/" + BookKeeperConstants.COOKIE_NODE
-                        + "/" + bookieId;
-        zkc.delete(bookieCookiePath, -1);
+        if (sv!=null) {
+            String bookieId = BookieImpl.getBookieId(sv).toString();
+            String bookieCookiePath =
+                    ZKMetadataDriverBase.resolveZkLedgersRootPath(sv)
+                            + "/" + BookKeeperConstants.COOKIE_NODE
+                            + "/" + bookieId;
+            zkc.delete(bookieCookiePath, -1);
+        try {
+                 result = BookKeeperAdmin.initBookie(sv);
+            } catch (Exception e) {
+                result =false;
 
-        boolean result = BookKeeperAdmin.initBookie(sv);
+            }
+
+
+        }
+
 
         Assert.assertEquals(expectedResult, result);
         System.out.println("test finished");
